backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
qdrant-client==1.7.1
langgraph==0.0.40
langchain==0.1.9
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
scikit-learn==1.3.2
google-generativeai==0.3.0
